fix: move Homebrew installation to extraUserActivation
nix-darwin's Homebrew module runs `brew bundle` as part of the user
activation, which comes before the system activation script. Thus,
on a fresh install, Homebrew was previously not yet available at that
point in the activation, causing it to fail. This change moves Homebrew
installation to an earlier point in the activation.